requirements.txt
Django>=5.1.7
python-dotenv>=1.0.0
razorpay>=1.4.2
Pillow>=10.0.0 